synchron Feature #28523
git-svn-id: svn+ssh://10.1.90.223/SVNroot/astra/server/trunk@193529 e9a6705e-e829-0410-83f2-cebdff2db7e5

+++ src/aodb.cc
@@ -2116,7 +2116,7 @@ void recNoFltNext( int point_id, std::string point_addr )
   Qry.Execute();
 }
 
-void setDelete( int aodb_point_id )
+void setDelete( double aodb_point_id )
 {
   TQuery Qry( &OraSession );
   Qry.SQLText =
@@ -2135,7 +2135,7 @@ bool isDelete( int point_id )
   return ( !Qry.Eof && Qry.FieldAsInteger( "pr_del" ) );
 }
 
-void setSCD_OUT( int aodb_point_id, TDateTime aodb_scd_out )
+void setSCD_OUT( double aodb_point_id, TDateTime aodb_scd_out )
 {
   TQuery Qry( &OraSession );
   Qry.SQLText =
